修改强制关闭线程机制
改为每个线程都发送信号到 record，超过 20 分钟信号不更新，则认为该线程处于僵死状态，强制关闭线程
