+++ ziyan/lib/Watchdog.py
@@ -1,5 +1,7 @@
 # -*- coding:utf-8 -*-
 
+import ctypes
+import inspect
 import threading
 import time
 
@@ -33,7 +35,8 @@ def watchdog(*args):
             threads_set = dict()
 
             for thread in dead_threads:
-                worker = threading.Thread(target=thread.work, args=(args[2],), kwargs={},
+                worker = threading.Thread(target=thread.work, args=(args[2],),
+                                          kwargs={'name': thread.name, 'record': args[3]},
                                           name='%s' % thread.name, daemon=True)
                 worker.start()
                 threads_set[thread.name] = worker
@@ -41,3 +44,23 @@ def watchdog(*args):
             args[3].thread_set = threads_set
 
         time.sleep(10)
+
+
+class Maintainer:
+    def __init__(self):
+        self.thread_signal = dict()
+        self.thread_set = None
+
+    def _async_raise(self, tid, exctype):
+        """raises the exception, performs cleanup if needed"""
+        tid = ctypes.c_long(tid)
+        if not inspect.isclass(exctype):
+            exctype = type(exctype)
+        res = ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, ctypes.py_object(exctype))
+        if res == 0:
+            log.error("invalid thread id")
+        elif res != 1:
+            """if it returns a number greater than one, you're in trouble,
+            and you should call it again with exc=NULL to revert the effect"""
+            ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, None)
+            log.error("PyThreadState_SetAsyncExc failed")
