+++ ziyan/script/manage.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import argparse
-import ctypes
-import datetime
 import glob
-import inspect
 import os
 import sys
 import time
@@ -15,14 +12,10 @@ from threading import Thread
 from plugins.your_plugin import *
 
 from ziyan.lib.Sender import Sender
-from ziyan.lib.Watchdog import watchdog
+from ziyan.lib.Watchdog import watchdog, Maintainer
 from ziyan.tests import Test_conf
 from ziyan.utils.logbook_wrapper import setup_logger
 from ziyan.utils.util import get_conf
-
-
-class Maintainer:
-    pass
 
 
 def start():
@@ -59,36 +52,23 @@ def start():
 
     thread_set = dict()
 
+    recorder = Maintainer()
+
     for worker in workers:
-        thread = Thread(target=worker.work, args=(queue,), kwargs={},
+        thread = Thread(target=worker.work, args=(queue,),
+                        kwargs={'name': worker.name, 'record': recorder},
                         name='%s' % worker.name, daemon=True)
         thread.start()
         thread_set[worker.name] = thread
 
-    recorder = Maintainer()
     recorder.thread_set = thread_set
-    recorder.date = datetime.date.today()
+
     Thread(target=watchdog, name='watchdog', args=(thread_set, workers, queue, recorder), daemon=True).start()
     return recorder
 
 
 def test():
     unittest.main(Test_conf, argv=sys.argv[1:])
-
-
-def _async_raise(tid, exctype):
-    """raises the exception, performs cleanup if needed"""
-    tid = ctypes.c_long(tid)
-    if not inspect.isclass(exctype):
-        exctype = type(exctype)
-    res = ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, ctypes.py_object(exctype))
-    if res == 0:
-        raise ValueError("invalid thread id")
-    elif res != 1:
-        """if it returns a number greater than one, you're in trouble,
-        and you should call it again with exc=NULL to revert the effect"""
-        ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, None)
-        raise SystemError("PyThreadState_SetAsyncExc failed")
 
 
 if __name__ == '__main__':
@@ -98,10 +78,9 @@ if __name__ == '__main__':
     if command == 'run':
         record = start()
         while True:
-            if record.date != datetime.date.today():
-                record.date = datetime.date.today()
-                for thread in record.thread_set.values():
-                    _async_raise(thread.ident, SystemExit)
+            for threadname, signal in record.thread_signal.items():
+                if time.time() - signal > 1200:
+                    record._async_raise(record.thread_set[threadname].ident, SystemExit)
             time.sleep(5)
     elif command == 'test':
         test()
